Percolation algorithm on BA network is implemented. And will be modified in the future for any network. Later Explosive percolation on on BA network will be implemented. Explosive percolation is and extension of Percolation, where there are sum rule and product rule.

+++ src/nets/link.cpp
@@ -8,7 +8,7 @@
 
 using namespace std;
 
-Link::Link(size_t id_a, size_t id_b) {
+Link::Link(uint id_a, uint id_b) {
     if(id_a == id_b){
         cout << "between " << id_a << " and " << id_b << endl;
         cerr << "Self linking is not allowed : line " << __LINE__ << endl;
