+++ src/nets/link.h
@@ -12,14 +12,14 @@
  *
  */
 class Link{
-    //size_t _id;
+    //uint _id;
     int _group_id{-1}; // will be used for percolation
     bool _active{false};
-    size_t _id_node_a, _id_node_b;
+    uint _id_node_a, _id_node_b;
 public:
     ~Link() = default;
     Link() = default;
-    Link(size_t id_a, size_t id_b);
+    Link(uint id_a, uint id_b);
 
     int get_group_id() const { return  _group_id;}
     void set_group_id(int id) {_group_id = id;}
@@ -33,8 +33,8 @@ public:
         return ss.str();
     }
 
-    size_t get_a() const {return _id_node_a;}
-    size_t get_b() const {return _id_node_b;}
+    uint get_a() const {return _id_node_a;}
+    uint get_b() const {return _id_node_b;}
 };
 
 std::ostream& operator<<(std::ostream& os, const Link& lnk);
