+++ src/nets/node.cpp
@@ -8,7 +8,7 @@
 using namespace std;
 
 
-Node::Node(size_t id) {
+Node::Node(uint id) {
     _id = id;
 }
 
@@ -16,13 +16,13 @@ void Node::add_neighbor(Node n) {
     _connected_nodes.push_back(n.get_id());
 }
 
-void Node::add_neighbor(size_t n) {
+void Node::add_neighbor(uint n) {
     _connected_nodes.push_back(n);
 }
 
 void Node::add_neighbor_checked(Node n) {
-    size_t i = n.get_id();
-    for(size_t k{}; k < _connected_nodes.size(); ++k){
+    uint i = n.get_id();
+    for(uint k{}; k < _connected_nodes.size(); ++k){
         if(i == _connected_nodes[k]){
             cerr << "Already a member : line " << __LINE__ << endl;
         }
@@ -31,8 +31,8 @@ void Node::add_neighbor_checked(Node n) {
 
 }
 
-void Node::add_neighbor_checked(size_t n) {
-    for(size_t k{}; k < _connected_nodes.size(); ++k){
+void Node::add_neighbor_checked(uint n) {
+    for(uint k{}; k < _connected_nodes.size(); ++k){
         if(n == _connected_nodes[k]){
             cerr << "Already a member : line " << __LINE__ << endl;
         }
